+++ include/gui/rectangle.h
@@ -7,10 +7,10 @@
 namespace gui {
   class Rectangle : public Widget {
     public:
-      Rectangle(WidgetCollection *parent, i32 x, i32 y, i8 w, i8 h, u8 r, u8 g, u8 b);
+      Rectangle(WidgetCollection *parent, i16 x, i16 y, i8 w, i8 h, u8 r, u8 g, u8 b);
       ~Rectangle();
       void OnSetup();
-      void OnMouseMove(i32 oldX, i32 oldY, i32 x, i32 y);
+      void OnMouseMove(i16 oldX, i16 oldY, i16 x, i16 y);
   };
 }
